+++ src/basherServiceProvider.php
@@ -18,7 +18,7 @@ class basherServiceProvider extends ServiceProvider
         'fitluismacedo\basher\Commands\Push',
         'fitluismacedo\basher\Commands\Revert',
         'fitluismacedo\basher\Commands\Tag',
-        'fitluismacedo\basher\Commands\Ignore',
+        'fitluismacedo\basher\Commands\File',
     ];
     /**
      * Perform post-registration booting of services.